+++ Projects/Addons/Sources/Navigation/ChangeAddonNavigation.swift
@@ -2,7 +2,6 @@ import Foundation
 import SwiftUI
 import hCore
 import hCoreUI
-import hGraphQL
 
 public struct ChangeAddonInput: Identifiable, Equatable {
     public var id: String = UUID().uuidString
@@ -26,13 +25,6 @@ public enum AddonSource: String, Codable {
     case insurances = "INSURANCES"
     case travelCertificates = "TRAVEL_CERTIFICATES"
     case deeplink = "DEEPLINK"
-
-    public var getSource: OctopusGraphQL.UpsellTravelAddonFlow {
-        switch self {
-        case .insurances: return .appOnlyUpsale
-        case .travelCertificates, .deeplink: return .appUpsellUpgrade
-        }
-    }
 }
 
 struct AddonInfo: Equatable, Identifiable {
