+++ Projects/Addons/Sources/Views/ChangeAddonViewModel.swift
@@ -1,7 +1,6 @@
 import SwiftUI
 import hCore
 import hCoreUI
-import hGraphQL
 
 @MainActor
 public class ChangeAddonViewModel: ObservableObject {
