+++ Projects/ChangeTier/Sources/Views/ChangeTierViewModel.swift
@@ -2,7 +2,6 @@ import Foundation
 import SwiftUI
 import hCore
 import hCoreUI
-import hGraphQL
 
 @MainActor
 public class ChangeTierViewModel: ObservableObject {
