+++ Projects/Contracts/Sources/Service/Protocols/ContractsClient.swift
@@ -12,4 +12,10 @@ public struct ContractsStack: Sendable {
     public let activeContracts: [Contract]
     public let pendingContracts: [Contract]
     public let terminatedContracts: [Contract]
+
+    public init(activeContracts: [Contract], pendingContracts: [Contract], terminatedContracts: [Contract]) {
+        self.activeContracts = activeContracts
+        self.pendingContracts = pendingContracts
+        self.terminatedContracts = terminatedContracts
+    }
 }
