+++ Projects/Contracts/Sources/View/ContractDocuments.swift
@@ -6,7 +6,6 @@ import SwiftUI
 import TerminateContracts
 import hCore
 import hCoreUI
-import hGraphQL
 
 struct ContractDocumentsView: View {
     @PresentableStore var contractStore: ContractStore
