+++ Projects/Contracts/Sources/View/CrossSellingItem.swift
@@ -4,7 +4,6 @@ import PresentableStore
 import SwiftUI
 import hCore
 import hCoreUI
-import hGraphQL
 
 struct CrossSellingItem: View {
     @PresentableStore var store: ContractStore
