+++ Projects/TravelCertificate/Sources/Models/TravelInsuranceContractSpecification.swift
@@ -12,7 +12,7 @@ public struct TravelInsuranceContractSpecification: Codable, Equatable, Hashable
     let email: String?
     let fullName: String
 
-    init(
+    public init(
         contractId: String,
         minStartDate: Date,
         maxStartDate: Date,
